Factored out SpinnerAdapter class
